# Inline Enum.into with bitstring implementation