--- a/lib/elixir/lib/collectable.ex
+++ b/lib/elixir/lib/collectable.ex
@@ -89,9 +89,12 @@ end
 
 defimpl Collectable, for: BitString do
   def into(original) do
-    {original, fn
-      acc, {:cont, x} when is_bitstring(x) -> [acc | x]
-      acc, :done -> IO.iodata_to_binary(acc)
+    {[original], fn
+      list, {:cont, x} -> [x | list]
+      list, :done ->
+        list
+        |> :lists.reverse()
+        |> IO.iodata_to_binary()
       _, :halt -> :ok
     end}
   end