@ericmj 2017-03-06
@tsubery Can you include benchmarks of the new implementation since it differs from your reference in the comparisons? I ran a few basic tests myself and could not get a difference significant enough to convince me the new implementation is more performant.