@tsubery 2017-03-06
Your are correct, the benchmark did not reflect the code performance. This is because ```into: []``` is actually short circuited and bypasses Collectable protocol. I added another fix and a more comprehensive benchmark to the PR.
The ~20% improvement of the Collectable implementation is arguably insignificant. Seems harmless improvement since it is doesn't add complexity.