@josevalim 2017-03-06
@tsubery can we keep only the second commit please? I still don't see any noticiable difference for the collectable one (not even 20%).