@tsubery 2017-03-06
These are the results I get with [this benchmark](https://gist.github.com/tsubery/303dbdbcd65f8de49b197029e3265978)
Seems like 50% slower on the bigger lists. Maybe it has to do with GC pressure.
I'm curious if it is something with my machine or the bench setup.

##### With input 100m #####
Name                       ips        average  deviation         median
skip collectable          0.97         1.03 s    ±23.54%         0.98 s
before                  0.0831        12.04 s     ±0.00%        12.04 s
flat_list               0.0551        18.16 s     ±0.00%        18.16 s

Comparison:
skip collectable          0.97
before                  0.0831 - 11.72x slower
flat_list               0.0551 - 17.68x slower

##### With input 1k #####
Name                       ips        average  deviation         median
skip collectable      108.83 K        9.19 μs   ±281.46%        9.00 μs
before                 20.33 K       49.20 μs   ±102.15%       40.00 μs
flat_list              17.23 K       58.03 μs  ±2336.83%       56.00 μs

Comparison:
skip collectable      108.83 K
before                 20.33 K - 5.35x slower
flat_list              17.23 K - 6.32x slower

##### With input 1m #####
Name                       ips        average  deviation         median
skip collectable        113.70        8.79 ms     ±3.59%        8.74 ms
flat_list                15.97       62.62 ms    ±21.45%       53.85 ms
before                   14.84       67.40 ms    ±31.98%       55.43 ms

Comparison:
skip collectable        113.70
flat_list                15.97 - 7.12x slower
before                   14.84 - 7.66x slower