@josevalim 2017-03-07
@tsubery are the benchmarks warming up the code before running it?