@josevalim 2017-03-07
@tsubery I need to investigate because I can't see those changes locally. But I think the PR is currently good as is, so let's at least merge the current inlining!

Thank you!