@tsubery 2017-03-07
Created a [repo](https://github.com/tsubery/misc_bench/) with the whole project to lock down benchee version if that helps investigation