@michalmuskala 2017-03-07
An important thing to notice is that this changes the behaviour of `Enum.into/2` with a binary.

Before `Enum.into([1], "")` would fail, now it will produce `<<1>>`