@tsubery 2017-03-07
@michalmuskala That's a good point. Both behaviours seem reasonable to me as long as it's consistent between Enum.into and Collectable implementations. Having said that, this commit might introduce some unexpected behavior.