@josevalim 2017-03-07
@tsubery we need to stick with the behaviour of `for ..., into: ""` which is optimized by the VM and only allow binaries, so we need to fix the behaviour @michalmuskala commented.