@tsubery 2017-03-07
Having another pass over the enumerable to verify all elements are binaries takes away most of the performance advantage introduced by this PR.